Moved the code under `benchmarks`
